osd: no need to remove mapping for non-snap objects

+++ src/osd/PG.cc
@@ -8509,10 +8509,12 @@ void PG::with_heartbeat_peers(std::function<void(int)> f)
 
 void PG::pg_remove_object(const ghobject_t& oid, ObjectStore::Transaction *t)
 {
-  OSDriver::OSTransaction _t(osdriver.get_transaction(t));
-  int r = snap_mapper.remove_oid(oid.hobj, &_t);
-  if (r != 0 && r != -ENOENT) {
-    ceph_abort();
+  if (oid.hobj.snap < CEPH_MAXSNAP) {
+    OSDriver::OSTransaction _t(osdriver.get_transaction(t));
+    int r = snap_mapper.remove_oid(oid.hobj, &_t);
+    if (r != 0 && r != -ENOENT) {
+      ceph_abort();
+    }
   }
   t->remove(coll, oid);
 }
